requirements.txt
requests
pyyaml
python-dotenv
openai
google-genai